Move mutations and queries to separate files
