+++ app/javascript/src/graphql/mutations/findRecipes.js
@@ -1,0 +1,21 @@
+import { gql } from '@apollo/client';
+
+export const FIND_RECIPES = gql`
+  mutation FindRecipes($input: FindRecipesInput!) {
+    findRecipes(input: $input) {
+      recipes {
+        id
+        title
+        ingredientIds
+        cookTime
+        prepTime
+        ratings
+        cuisine
+        category
+        author
+        image
+        matchingIngredientsCount
+      }
+    }
+  }
+`;
