+++ app/javascript/src/graphql/queries/getIngredients.js
@@ -1,0 +1,10 @@
+import { gql } from '@apollo/client';
+
+export const GET_INGREDIENTS = gql`
+  query GetIngredients {
+    ingredients {
+      id
+      name
+    }
+  }
+`;
